+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 [Unreleased]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.2...HEAD
 
+### Added
+
+ - Uniform buffer object functionality.
+
 ## [0.4.2] - 2019-04-10
 [0.4.2]: https://github.com/mickvangelderen/gl-typed-rust/compare/v0.4.1...0.4.2